♻ refactor: 音関係をuseSound.txに切り出し
